Added Documentation Moved FramAPIImpl to separate class :)
